+++ .github/workflows/docs/conf.py
@@ -21,13 +21,15 @@ html_theme_options = {
     "repository_url": "https://github.com/CQCL/pytket-qsharp",
     "use_repository_button": True,
     "use_issues_button": True,
+    "logo": {
+        "image_light": "Quantinuum_logo_black.png",
+        "image_dark": "Quantinuum_logo_white.png",
+    },
 }
 
 html_static_path = ["_static"]
 
 html_css_files = ["custom.css"]
-
-html_logo = "Quantinuum_logo.png"
 
 # -- Extension configuration -------------------------------------------------
 
@@ -114,7 +116,6 @@ def correct_signature(
     signature: str,
     return_annotation: str,
 ) -> (str, str):
-
     new_signature = signature
     new_return_annotation = return_annotation
     for k, v in app.config.custom_internal_mapping.items():