+++ pytket/extensions/qsharp/backends/common.py
@@ -134,7 +134,6 @@ class _QsharpBaseBackend(Backend):
     def default_compilation_pass(self, optimisation_level: int = 2) -> BasePass:
         assert optimisation_level in range(3)
         if optimisation_level == 0:
-
             return SequencePass(
                 [DecomposeBoxes(), FlattenRegisters(), self.rebase_pass()]
             )